Fix bug for linux compile time

+++ kabirrec/surprise/prediction_algorithms/weighted_slope_one.pyx
@@ -37,7 +37,7 @@ class WeightedSlopeOne(AlgoBase):
 
         AlgoBase.fit(self, trainset)
 
-        freq = np.zeros((trainset.n_items, trainset.n_items), np.int32)
+        freq = np.zeros((trainset.n_items, trainset.n_items), np.int_)
         dev = np.zeros((trainset.n_items, trainset.n_items), np.double)
 
         # Computation of freq and dev arrays.
